requirements.txt
Package           Version
----------------- -----------
distlib           0.3.9
et_xmlfile        2.0.0
filelock          3.16.1
fsspec            2024.10.0
Jinja2            3.1.4
MarkupSafe        3.0.2
mpmath            1.3.0
networkx          3.4.2
numpy             2.2.0
openpyxl          3.1.5
pandas            2.2.3
pillow            11.0.0
pip               24.3.1
platformdirs      4.3.6
PyMySQL           1.1.1
python-dateutil   2.9.0.post0
pytz              2024.2
setuptools        65.5.0
six               1.17.0
sympy             1.13.1
torch             2.5.1
torchvision       0.20.1
typing_extensions 4.12.2
tzdata            2024.2
virtualenv        20.28.0
watchdog          6.0.0
